requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-pinecone
pinecone-client # new added
pinecone
pypdf
python-dotenv
google-genai # new -added 